refactor: Game 구조체 생성 및 기존 코드 Game 타입 내부로 이동
