--- RockPaperScissors/RockPaperScissors/enums.swift
@@ -1,21 +1,0 @@
-//
-//  enums.swift
-//  RockPaperScissors
-//
-//  Created by 전민수 on 2022/04/19.
-//
-
-import Foundation
-
-enum RockPaperScissors: Int {
-    case ready = 0
-    case scissors = 1
-    case rock = 2
-    case paper = 3
-}
-
-enum GameResult {
-    case usersVictory
-    case computersVictory
-    case tie
-}
